# Refactor integer parsing, add string to 128bit integer methods