@BlobCodes 2021-08-13
Note that some checks won't run successfully because of the bad 128bit support (complaining about missing __modti3, __divti3, __udivti3, __umodti3).