@asterite 2021-08-13
> https://forum.crystal-lang.org/t/1-0-multi-threading-memory-use-issues/2454/15?u=asterite

Yeah, I actually wish we removed supporting 128 bit integers. Now I have to implement those in the interpreter too, sigh... :-(

Very few languages support 128 bits. I wish I didn't implement it back then...