@asterite 2021-08-13
@BlobCodes Of course my comment has nothing to do with your PR, which is very welcome! 🙏  But I'm not sure 128 bits support will be easy to implement...