@HertzDevil 2021-08-14
The added 128-bit methods are completely independent from the performance optimizations. These two should belong in separate PRs.